restructure former indentationlevelparser #660
