+++ analysis/import/IndentationLevelParser/src/main/kotlin/de/maibornwolff/codecharta/importer/indentationlevelparser/ProjectGenerator.kt
@@ -1,6 +1,7 @@
 package de.maibornwolff.codecharta.importer.indentationlevelparser
 
 import de.maibornwolff.codecharta.filter.mergefilter.MergeFilter
+import de.maibornwolff.codecharta.importer.indentationlevelparser.model.FileMetrics
 import de.maibornwolff.codecharta.model.MutableNode
 import de.maibornwolff.codecharta.model.PathFactory
 import de.maibornwolff.codecharta.model.Project
